stt-service/requirements.txt
faster-whisper
silero-vad
sounddevice
pyzmq
numpy
